requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
gspread==6.0.0
google-auth==2.27.0
google-genai==0.3.2
python-dateutil==2.8.2